+++ src/main/java/koreatech/in/service/OwnerServiceImpl.java
@@ -30,7 +30,6 @@ import koreatech.in.util.SesMailSender;
 import koreatech.in.util.SlackNotiSender;
 import koreatech.in.util.StringRedisUtilObj;
 import koreatech.in.util.StringRedisUtilStr;
-import koreatech.in.util.jwt.TemporaryAccessJwtGenerator;
 import org.apache.velocity.app.VelocityEngine;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
