+++ src/main/java/koreatech/in/service/admin/AdminUserServiceImpl.java
@@ -71,8 +71,8 @@ import koreatech.in.repository.user.UserMapper;
 import koreatech.in.service.JwtValidator;
 import koreatech.in.service.RefreshJwtValidator;
 import koreatech.in.util.StringRedisUtilObj;
-import koreatech.in.util.jwt.UserAccessJwtGenerator;
-import koreatech.in.util.jwt.UserRefreshJwtGenerator;
+import koreatech.in.service.UserAccessJwtGenerator;
+import koreatech.in.service.UserRefreshJwtGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
